App Catalog Review Fixes
