+++ samples/graph-appcatalog-lifecycle/csharp/AppCatalogSample/appsettings.json
@@ -1,5 +1,5 @@
 {
-"MicrosoftAppId": "",
-"MicrosoftAppPassword": "",
-"ConnectionName": ""
+  "MicrosoftAppId": "",
+  "MicrosoftAppPassword": "",
+  "ConnectionName": ""
 }
